Update good-touch-principle-question-4.md
The sentence "Green 3 needs be played on the stacks." is misleading: it suggests that the green 3 HAS to be clued right now. I think the example is better with this sentence simply removed. (If you disagree, then the sentence needs to be gramatically fixed to "Green 3 needs to be played on the stacks.")